Use new error type from main
All of the old error types can now be converted losslessly into the new
type, and this is using the new pretty-printer to emit and color the
errors. The next steps are deleting the old errors and replacing them
with the new ones one by one.
