+++ golden/error/lex_unclosed_string_literal.test
@@ -3,7 +3,7 @@
 # output:
 stdin:2:1
   ╷
-2 │ 
+2 │
   ╵ ^
 Error: Unexpected end of input, string literal is not closed.
