+++ golden/error/lex_unclosed_triple_string_literal.test
@@ -4,7 +4,7 @@ This is unclosed as well
 # output:
 stdin:3:1
   ╷
-3 │ 
+3 │
   ╵ ^
 Error: Unexpected end of input, string literal is not closed.
