+++ golden/error/lex_unmatched_lbracket_eof.test
@@ -3,7 +3,7 @@
 # output:
 stdin:2:1
   ╷
-2 │ 
+2 │
   ╵ ^
 Error: Expected ']'.
