+++ src/error_old.rs
@@ -9,7 +9,7 @@
 
 use std::rc::Rc;
 
-use crate::source::{Inputs, Span};
+use crate::source::Span;
 
 pub type Result<T> = std::result::Result<T, Box<dyn Error>>;
 
@@ -31,139 +31,6 @@ pub trait Error: std::fmt::Debug {
 
     /// Optionally, additional information, or a hint on how to fix the problem.
     fn help(&self) -> Option<&str>;
-}
-
-impl dyn Error {
-    pub fn print(&self, inputs: &Inputs) {
-        let bold_red = "\x1b[31;1m";
-        let bold_yellow = "\x1b[33;1m";
-        let reset = "\x1b[0m";
-
-        if let Some(span) = self.span() {
-            let highlight = highlight_span_in_line(inputs, span, bold_red);
-            eprint!("{}", highlight);
-            eprintln!("{}Error:{} {}", bold_red, reset, self.message());
-        } else {
-            eprintln!("{}Error:{} {}", bold_red, reset, self.message());
-        }
-
-        for (note_span, note) in self.notes() {
-            let highlight = highlight_span_in_line(inputs, *note_span, bold_yellow);
-            eprint!("\n{}", highlight);
-            eprintln!("{}Note:{} {}", bold_yellow, reset, note);
-        }
-
-        if let Some(help) = self.help() {
-            eprintln!("\n{}Help:{} {}", bold_yellow, reset, help);
-        }
-    }
-}
-
-fn highlight_span_in_line(inputs: &Inputs, span: Span, highlight_ansi: &str) -> String {
-    use std::cmp;
-    use std::fmt::Write;
-    use unicode_width::UnicodeWidthStr;
-
-    let doc = &inputs[span.doc().0 as usize];
-    let input = doc.data;
-
-    // Locate the line that contains the error.
-    let mut line = 1;
-    let mut line_start = 0;
-    let mut line_end = 0;
-    for (&c, i) in input.as_bytes().iter().zip(0..) {
-        if i == span.start() {
-            break;
-        }
-        if c == b'\n' {
-            line += 1;
-            line_start = i + 1;
-        }
-    }
-    for (&c, i) in input.as_bytes()[line_start..].iter().zip(line_start..) {
-        if c == b'\n' {
-            line_end = i;
-            break;
-        }
-    }
-    if line_end <= line_start {
-        line_end = input.len();
-    }
-
-    // Save this for reporting the error location, in case we adjust the line
-    // start below. Add 1 because the first column is column 1, not 0.
-    let column = 1 + span.start() - line_start;
-
-    // If there is a really long line (for example, because you are evaluating
-    // a multi-megabyte json document that is formatted without whitespace, on
-    // a single line), then printing the error to the terminal is going to line
-    // wrap that enormous content and not do anything productive. In that case,
-    // we would rather just truncate. If we do truncate, add an ellipsis to
-    // clarify that we did.
-    let mut trunc_prefix = "";
-    let mut trunc_suffix = "";
-    if span.start() - line_start > 100 {
-        line_start = span.start() - 20;
-        // Ensure we don't slice code points in half. Slightly nicer would be
-        // to not slice grapheme clusters in half (and also measure whether the
-        // line is long from its width, not in bytes), but that is way overkill
-        // for a fringe case like this.
-        while !input.is_char_boundary(line_start) {
-            line_start -= 1;
-        }
-        trunc_prefix = "…";
-    }
-    if line_end > span.end() + 100 {
-        line_end = span.end() + 20;
-        while !input.is_char_boundary(line_end) {
-            line_end += 1;
-        }
-        trunc_suffix = "…";
-    }
-
-    let line_content = &input[line_start..line_end];
-
-    // The length of the mark can be longer than the line, for example when
-    // token to mark was a multiline string literal. In that case, highlight
-    // only up to the newline, don't extend the tildes too far.
-    let indent_content = &line_content[..span.start() - line_start];
-    let as_of_error = &line_content[span.start() - line_start..];
-    let error_content = &as_of_error[..cmp::min(span.len(), as_of_error.len())];
-
-    // The width of the error is not necessarily the number of bytes,
-    // measure the Unicode width of the span to underline.
-    let indent_width = indent_content.width() + trunc_prefix.width();
-    let mark_width = cmp::max(1, error_content.width());
-
-    let line_num_str = line.to_string();
-    let line_num_pad: String = line_num_str.chars().map(|_| ' ').collect();
-    let mark_indent: String = " ".repeat(indent_width);
-    let mark_under: String = "~".repeat(mark_width);
-
-    let reset = "\x1b[0m";
-
-    let mut result = String::new();
-    // Note, the unwraps here are safe because writing to a string does not fail.
-    writeln!(&mut result, "{}:{}:{}", doc.name, line, column,).unwrap();
-    writeln!(&mut result, "{} {}╷{}", line_num_pad, highlight_ansi, reset).unwrap();
-    writeln!(
-        &mut result,
-        "{} {}│{} {}{}{}",
-        line_num_str, highlight_ansi, reset, trunc_prefix, line_content, trunc_suffix
-    )
-    .unwrap();
-    writeln!(
-        &mut result,
-        "{} {}╵ {}^{}{}",
-        line_num_pad,
-        highlight_ansi,
-        mark_indent,
-        &mark_under[1..],
-        reset
-    )
-    .unwrap();
-
-    result
 }
 
 /// We encountered some IO failure.
